update logging for webservice config

+++ src/lib/core/domain/config/webserver_jwt_access_token_expiry_secs.rs
@@ -3,7 +3,7 @@ use thiserror::Error;
 
 /// A validated and formatted Db connection retry initial delay in seconds.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct WebserverJwtAccessTokenExpirySecs(u32);
+pub struct WebserverJwtAccessTokenExpirySecs(u64);
 
 impl fmt::Display for WebserverJwtAccessTokenExpirySecs {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -16,7 +16,7 @@ pub enum WebserverJwtAccessTokenExpirySecsError {}
 
 impl WebserverJwtAccessTokenExpirySecs {
     pub fn new(
-        raw_jwt_access_token_expiry_secs: &u32,
+        raw_jwt_access_token_expiry_secs: &u64,
     ) -> Result<Self, WebserverJwtAccessTokenExpirySecsError> {
         // Add validation logic here if needed
         Ok(WebserverJwtAccessTokenExpirySecs(
@@ -26,7 +26,7 @@ impl WebserverJwtAccessTokenExpirySecs {
     pub fn default() -> Self {
         WebserverJwtAccessTokenExpirySecs(900)
     }
-    pub fn get(&self) -> u32 {
+    pub fn get(&self) -> u64 {
         self.0
     }
 }
