+++ src/lib/core/domain/config/webserver_shutdown_timeout_secs.rs
@@ -3,7 +3,7 @@ use thiserror::Error;
 
 /// A validated and formatted Db connection retry initial delay in seconds.
 #[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct WebserverShutdownTimeoutSecs(u32);
+pub struct WebserverShutdownTimeoutSecs(u64);
 
 impl fmt::Display for WebserverShutdownTimeoutSecs {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
@@ -15,14 +15,14 @@ impl fmt::Display for WebserverShutdownTimeoutSecs {
 pub enum WebserverShutdownTimeoutSecsError {}
 
 impl WebserverShutdownTimeoutSecs {
-    pub fn new(raw_shutdown_timeout_secs: &u32) -> Result<Self, WebserverShutdownTimeoutSecsError> {
+    pub fn new(raw_shutdown_timeout_secs: &u64) -> Result<Self, WebserverShutdownTimeoutSecsError> {
         // Add validation logic here if needed
         Ok(WebserverShutdownTimeoutSecs(*raw_shutdown_timeout_secs))
     }
     pub fn default() -> Self {
         WebserverShutdownTimeoutSecs(300)
     }
-    pub fn get(&self) -> u32 {
+    pub fn get(&self) -> u64 {
         self.0
     }
 }
